refactor(arch): Merge permissions into verify_token
